requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
cryptography==41.0.7
distlib==0.3.8
exceptiongroup==1.2.0
fastapi==0.105.0
fastapi-pagination==0.12.14
filelock==3.13.1
h11==0.14.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.1
openopc2==0.1.11
peewee==3.17.0
platformdirs==4.1.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.8.0
Pyro5==5.15
python-dotenv==1.0.0
requests==2.31.0
rich==12.6.0
serpent==1.41
shellingham==1.5.4
sniffio==1.3.0
starlette==0.27.0
typer==0.6.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
virtualenv==20.25.0
